AI/requirements.txt
pandas
openai
langchain
langchain-community
langchain-openai
chromadb
tiktoken
faiss-cpu
jq
Flask
python-dotenv
scikit-learn
numpy
qdrant-client